+++ plone/app/content/browser/query.py
@@ -11,5 +11,7 @@ class QueryStringIndexOptions(BrowserView):
     def __call__(self):
         registry = getUtility(IRegistry)
         config = IQuerystringRegistryReader(registry)()
-        self.request.response.setHeader("Content-Type", "application/json")
+        self.request.response.setHeader(
+            'Content-Type', 'application/json; charset=utf-8'
+        )
         return json.dumps(config)